https://app.powerbi.com/links/zytFc6J1DI?ctid=bb42b609-7f4a-43d0-ba23-49a6ece5ce02&pbi_source=linkShare
Using the BikeTrip Dataset from Google BigQuery (bigquery-public-data.austin_bikeshare), I created a Power BI Dashboard to compare & analyse different aspects of the information provided in the dataset. This analysis allowed me to refine the data and create a Power BI dashboard that makes the general public's understanding of the data easier.